src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.4.0",
    "redis>=5.2.1",
    "httpx>=0.27.2",
    "requests>=2.32.3",
    "sqlmodel>=0.0.22",
    "tenacity>=9.0.0",
    "fastapi[cli,standard]>=0.115.0",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.5.2",
    "python-dotenv>=1.0.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
]
